+++ examples/pointer_impls/associated.rs
@@ -4,7 +4,7 @@
 //  Created:
 //    16 Dec 2024, 14:27:33
 //  Last edited:
-//    16 Dec 2024, 14:29:59
+//    17 Dec 2024, 16:38:22
 //  Auto updated?
 //    Yes
 //